Fix duplicated code in transaction address getters
