+++ src/Entity/Transaction/BroadcastTransaction.php
@@ -20,8 +20,6 @@
 
 namespace Adshares\Ads\Entity\Transaction;
 
-use Adshares\Ads\Util\AdsChecksumGenerator;
-
 /**
  * Transaction type=<'broadcast'>.
  *
@@ -29,6 +27,8 @@ use Adshares\Ads\Util\AdsChecksumGenerator;
  */
 class BroadcastTransaction extends AbstractTransaction
 {
+    use GetSenderAddressTrait;
+
     /**
      * @var string
      */
@@ -94,19 +94,6 @@ class BroadcastTransaction extends AbstractTransaction
     public function getNode(): int
     {
         return $this->node;
-    }
-
-    /**
-     * @return string
-     */
-    public function getSenderAddress(): string
-    {
-        return sprintf(
-            '%04X-%08X-%s',
-            $this->node,
-            $this->user,
-            AdsChecksumGenerator::getAccountChecksum($this->node, $this->user)
-        );
     }
 
     /**
