+++ src/Entity/Transaction/KeyTransaction.php
@@ -20,8 +20,6 @@
 
 namespace Adshares\Ads\Entity\Transaction;
 
-use Adshares\Ads\Util\AdsChecksumGenerator;
-
 /**
  * Transaction type=<'account_created', 'change_account_key', 'change_node_key'>.
  *
@@ -29,6 +27,9 @@ use Adshares\Ads\Util\AdsChecksumGenerator;
  */
 class KeyTransaction extends AbstractTransaction
 {
+    use GetSenderAddressTrait;
+    use GetTargetAddressTrait;
+
     /**
      * @var int
      */
@@ -133,41 +134,11 @@ class KeyTransaction extends AbstractTransaction
     }
 
     /**
-     * @return string
-     */
-    public function getSenderAddress(): string
-    {
-        return sprintf(
-            '%04X-%08X-%s',
-            $this->node,
-            $this->user,
-            AdsChecksumGenerator::getAccountChecksum($this->node, $this->user)
-        );
-    }
-
-    /**
      * @return null|string
      */
     public function getSignature(): ?string
     {
         return $this->signature;
-    }
-
-    /**
-     * @return null|string
-     */
-    public function getTargetAddress(): ?string
-    {
-        if (null === $this->targetNode || null === $this->targetUser) {
-            return null;
-        }
-
-        return sprintf(
-            '%04X-%08X-%s',
-            $this->targetNode,
-            $this->targetUser,
-            AdsChecksumGenerator::getAccountChecksum($this->targetNode, $this->targetUser)
-        );
     }
 
     /**
