+++ src/Entity/Transaction/LogAccountTransaction.php
@@ -21,7 +21,6 @@
 namespace Adshares\Ads\Entity\Transaction;
 
 use Adshares\Ads\Entity\Account;
-use Adshares\Ads\Util\AdsChecksumGenerator;
 
 /**
  * Transaction type=<'log_account'>.
@@ -30,6 +29,8 @@ use Adshares\Ads\Util\AdsChecksumGenerator;
  */
 class LogAccountTransaction extends AbstractTransaction
 {
+    use GetSenderAddressTrait;
+
     /**
      * @var int
      */
@@ -82,19 +83,6 @@ class LogAccountTransaction extends AbstractTransaction
     public function getNode(): int
     {
         return $this->node;
-    }
-
-    /**
-     * @return string
-     */
-    public function getSenderAddress(): string
-    {
-        return sprintf(
-            '%04X-%08X-%s',
-            $this->node,
-            $this->user,
-            AdsChecksumGenerator::getAccountChecksum($this->node, $this->user)
-        );
     }
 
     /**
